seeds: sync 2026-07-17
