+++ .seeds/issues.jsonl
@@ -976,7 +976,7 @@
 {"id":"warren-a0a2","title":"scheduler: bound transient spawn-failure retries — back off / consume slot after N attempts, stop minting a failed run row per tick","status":"closed","type":"task","priority":2,"createdAt":"2026-07-14T15:50:20.648Z","updatedAt":"2026-07-14T17:19:22.637Z","description":"dispatchCronTrigger treats any non-agent-not-found spawn failure as transient and retries every 60s tick within the same slot, and spawnRun creates the run row before contacting the runtime, so a persistently-unreachable runtime (e.g. the pre-warren-1ca1 K8s misconfig, or burrow down for hours) mints one failed/never_started run row per tick indefinitely (~200 rows in 3h observed on GKE). Harden: (1) cap in-slot retries or add exponential backoff per trigger, consuming the slot after N consecutive transient failures with a scheduler.cron_gave_up log; (2) consider not persisting (or GC-ing) never_started rows from scheduler retries so the runs list isn't flooded. Keep genuine brief-outage retry behavior (pl-2f15 risk #5) intact.","closedAt":"2026-07-14T17:19:22.637Z"}
 {"id":"warren-035c","title":"warren bot commits are not immune to inherited GIT_AUTHOR_*/GIT_COMMITTER_* env — breaks Article VII + fails clone-apply test under pre-commit hook","status":"closed","type":"bug","priority":2,"createdAt":"2026-07-14T15:59:31.485Z","updatedAt":"2026-07-14T20:40:53.450Z","description":"warrenCommitIdentityArgs() pins the bot identity via '-c user.name/user.email', but git's GIT_AUTHOR_NAME/GIT_AUTHOR_EMAIL env vars take precedence over config. Any warren process inheriting those vars mis-authors the reap bookkeeping commits (stage.ts x2, clone-apply.ts) and the plot-sync commit (plots/sync.ts) as the operator instead of 'warren <warren@os-eco.dev>'. Concrete repro: 'git commit' exports GIT_AUTHOR_* to its hooks, so this repo's own pre-commit (check:all -> bun test) deterministically fails clone-apply.test.ts 'commits as the warren bot' on every local commit; GIT_AUTHOR_NAME='x' bun test src/runs/reap/clone-apply.test.ts reproduces directly. Fix: add warrenCommitIdentityEnv() to src/bot-identity.ts (explicit GIT_AUTHOR_/GIT_COMMITTER_ name+email), extend ReapExec.run opts and projects/clone.ts SpawnFn options with optional env, and pass the identity env at the four commit sites so inherited env can never win.","closedAt":"2026-07-14T20:40:53.450Z"}
 {"id":"warren-39e1","title":"K8s /readyz: add k8s_api_reachable check via a PodWatcher isSynced()-style seam (positive K8s-topology readiness signal; scoped out of warren-c128)","status":"open","type":"task","priority":3,"createdAt":"2026-07-14T16:51:23.981Z","updatedAt":"2026-07-14T16:51:23.981Z","labels":["k8s-migration"]}
-{"id":"warren-9574","title":"Relocate authenticatedCloneUrl from src/runtime/k8s/workspace-init.ts to a neutral home (src/workspace/git/) — reap now imports it from an init-container module (pr-open.ts, finalize-collect.ts)","status":"open","type":"task","priority":3,"createdAt":"2026-07-14T17:29:56.845Z","updatedAt":"2026-07-14T17:29:56.845Z","labels":["k8s-migration"]}
+{"id":"warren-9574","title":"Relocate authenticatedCloneUrl from src/runtime/k8s/workspace-init.ts to a neutral home (src/workspace/git/) — reap now imports it from an init-container module (pr-open.ts, finalize-collect.ts)","status":"closed","type":"task","priority":3,"createdAt":"2026-07-14T17:29:56.845Z","updatedAt":"2026-07-17T04:55:03.728Z","labels":["k8s-migration"],"closedAt":"2026-07-17T04:55:03.728Z"}
 {"id":"warren-25d5","title":"Migrate warren-repo-cache PVC to an RWX storage class (GKE Filestore CSI) before adding a second node; pin storageClassName via overlay (deferred from warren-e908)","status":"open","type":"task","priority":3,"createdAt":"2026-07-14T17:57:57.558Z","updatedAt":"2026-07-14T17:57:57.558Z","labels":["k8s-migration"]}
 {"id":"warren-e36c","title":"README status pass: reconcile acceptance-scenario count and Fly.io-vs-GKE status/version line after k8s migration (flagged during warren-5b6d)","status":"closed","type":"task","priority":3,"createdAt":"2026-07-14T18:14:11.931Z","updatedAt":"2026-07-16T18:13:39.082Z","labels":["k8s-migration"],"plan_id":"pl-829f","description":"<!-- seeds:plan-backref:start -->\nAdopted into plan pl-829f.\n\nParent seed: warren-e176 — K8s migration design doc: pod-per-run isolation to replace single-Machine Fly co-tenancy\nPlan template: feature\nPlan approach: Introduce a RuntimeProvider contract (docs/design/runtime-provider-contract.md) that the warren domain depends on, with two backends behind it: LocalProvider (wraps today's burrow socket path — self-host, macOS, local dev) and K8sProvider…\n\nRun `sd plan show pl-829f` for the full plan (context, alternatives, sibling steps, acceptance criteria).\n<!-- seeds:plan-backref:end -->","blocks":["warren-e176","warren-5e45"],"closedAt":"2026-07-16T18:13:39.082Z"}
 {"id":"warren-23dd","title":"extend GIT_* repo-context scrub to stage.ts and plots/sync.ts warren-bot commit sites","status":"closed","type":"task","priority":3,"createdAt":"2026-07-14T20:40:53.561Z","updatedAt":"2026-07-14T20:52:53.602Z","description":"warren-fa84 fixed the GIT_DIR/GIT_INDEX_FILE env leak at the clone-apply spawn site via gitRepoContextScrubEnv() + resolveSpawnEnv() removal semantics (env values of undefined drop vars from the child env). The other warren-bot commit sites — both commits in src/runs/reap/stage.ts and the plot-sync commit in src/plots/sync.ts — still pass only env: warrenCommitIdentityEnv(), so they carry the identical latent leak if warren ever runs under a hook/parent that exports GIT_DIR. Compose the scrub there too: { ...gitRepoContextScrubEnv(), ...warrenCommitIdentityEnv() }, mirroring clone-apply.ts, and extend the fakeExec env assertions in stage.test.ts / sync.test.ts.","closedAt":"2026-07-14T20:52:53.602Z"}
