+++ src/runtime/k8s/workspace-init.test.ts
@@ -6,7 +6,6 @@ import { WorkspaceMaterializationError } from "../../workspace/errors.ts";
 import { cleanGitEnv } from "../../workspace/git/exec.test.ts";
 import { runGit } from "../../workspace/git/exec.ts";
 import {
-	authenticatedCloneUrl,
 	type InitFs,
 	type InitGitRunner,
 	mirrorPathFor,
@@ -110,28 +109,6 @@ describe("parseInitEnv", () => {
 		expect(() =>
 			parseInitEnv({ WARREN_REPO_URL: "  ", WARREN_BRANCH: "b", WARREN_BASE_BRANCH: "main" }),
 		).toThrow(/WARREN_REPO_URL/);
-	});
-});
-
-describe("authenticatedCloneUrl", () => {
-	test("injects x-access-token for https URLs", () => {
-		expect(authenticatedCloneUrl("https://github.com/o/r.git", "tok")).toBe(
-			"https://x-access-token:tok@github.com/o/r.git",
-		);
-	});
-
-	test("leaves the URL alone without a token", () => {
-		expect(authenticatedCloneUrl("https://github.com/o/r.git")).toBe("https://github.com/o/r.git");
-	});
-
-	test("does not touch ssh URLs", () => {
-		expect(authenticatedCloneUrl("git@github.com:o/r.git", "tok")).toBe("git@github.com:o/r.git");
-	});
-
-	test("does not double-inject when the authority already has userinfo", () => {
-		expect(authenticatedCloneUrl("https://user:pw@github.com/o/r.git", "tok")).toBe(
-			"https://user:pw@github.com/o/r.git",
-		);
 	});
 });
